+++ aiku/aiku-main/src/main/java/aiku_main/application_event/handler/BettingHandler.java
@@ -2,7 +2,7 @@ package aiku_main.application_event.handler;
 
 import aiku_main.application_event.event.ScheduleCloseEvent;
 import aiku_main.application_event.event.ScheduleExitEvent;
-import aiku_main.service.BettingService;
+import aiku_main.service.betting.BettingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
