+++ aiku/aiku-main/src/main/java/aiku_main/application_event/handler/PointChangeEventHandler.java
@@ -1,11 +1,7 @@
 package aiku_main.application_event.handler;
 
 import aiku_main.application_event.event.PointChangeEvent;
-import aiku_main.service.MemberPointChangeFacade;
-import aiku_main.service.MemberService;
-import aiku_main.service.PointLogFactory;
-import aiku_main.service.PointLogService;
-import common.domain.log.PointLog;
+import aiku_main.service.member.MemberPointChangeFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
