+++ aiku/aiku-main/src/main/java/aiku_main/application_event/handler/ScheduleHandler.java
@@ -4,7 +4,7 @@ import aiku_main.application_event.event.ScheduleAutoCloseEvent;
 import aiku_main.application_event.event.ScheduleCloseEvent;
 import aiku_main.application_event.event.ScheduleOpenEvent;
 import aiku_main.application_event.event.TeamExitEvent;
-import aiku_main.service.ScheduleService;
+import aiku_main.service.schedule.ScheduleService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
