+++ aiku/aiku-main/src/main/java/aiku_main/application_event/handler/TeamHandler.java
@@ -2,7 +2,7 @@ package aiku_main.application_event.handler;
 
 import aiku_main.application_event.event.ScheduleCloseEvent;
 import aiku_main.application_event.event.TeamExitEvent;
-import aiku_main.service.TeamService;
+import aiku_main.service.team.TeamService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
