+++ aiku/aiku-main/src/main/java/aiku_main/application_event/handler/TitleHandler.java
@@ -1,7 +1,7 @@
 package aiku_main.application_event.handler;
 
 import aiku_main.application_event.event.ScheduleCloseEvent;
-import aiku_main.service.TitleService;
+import aiku_main.service.title.TitleService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
