+++ aiku/aiku-main/src/main/java/aiku_main/controller/FirebaseController.java
@@ -1,9 +1,7 @@
 package aiku_main.controller;
 
 import aiku_main.dto.*;
-import aiku_main.service.LoginService;
-import aiku_main.service.MemberFirebaseService;
-import common.domain.member.Member;
+import aiku_main.service.member.MemberFirebaseService;
 import common.response.BaseResponse;
 import common.response.BaseResultDto;
 import jakarta.validation.Valid;
