+++ aiku/aiku-main/src/main/java/aiku_main/service/member/EmailService.java
@@ -1,4 +1,4 @@
-package aiku_main.service;
+package aiku_main.service.member;
 
 import aiku_main.dto.InquiryDto;
 import aiku_main.exception.MemberNotFoundException;
