+++ aiku/aiku-main/src/main/java/aiku_main/service/member/MemberRegisterService.java
@@ -1,4 +1,4 @@
-package aiku_main.service;
+package aiku_main.service.member;
 
 import aiku_main.application_event.event.PointChangeReason;
 import aiku_main.application_event.event.PointChangeType;
