+++ aiku/aiku-main/src/test/java/aiku_main/controller/BettingControllerTest.java
@@ -1,7 +1,7 @@
 package aiku_main.controller;
 
 import aiku_main.dto.betting.BettingAddDto;
-import aiku_main.service.BettingService;
+import aiku_main.service.betting.BettingService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.junit.jupiter.api.BeforeEach;
