+++ aiku/aiku-main/src/test/java/aiku_main/service/MemberFirebaseServiceTest.java
@@ -3,6 +3,7 @@ package aiku_main.service;
 import aiku_main.dto.FirebaseTokenDto;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.repository.member.MemberRepository;
+import aiku_main.service.member.MemberFirebaseService;
 import common.domain.member.Member;
 import common.domain.member.MemberProfileBackground;
 import common.domain.member.MemberProfileCharacter;
