+++ aiku/aiku-main/src/test/java/aiku_main/service/TermServiceTest.java
@@ -3,6 +3,7 @@ package aiku_main.service;
 import aiku_main.dto.member.TermResDto;
 import aiku_main.exception.TermException;
 import aiku_main.repository.term.TermRepository;
+import aiku_main.service.term.TermService;
 import common.domain.AgreedType;
 import common.domain.Term;
 import common.domain.TermTitle;
