+++ aiku/aiku-main/src/test/java/aiku_main/service/TitleServiceTest.java
@@ -1,5 +1,7 @@
 package aiku_main.service;
 
+import aiku_main.service.schedule.ScheduleService;
+import aiku_main.service.title.TitleService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.Betting;
 import common.domain.Location;
